Avoid relying on bundler internals
This logic uses code in whatever the default bundler installed by the
ruby version used by Heroku uses.

On the one hand, bundler does not really expose an "official" API for
doing this, so the results might not be fully consistent across bundler
versions and the entrypoint and API is not guaranteed to be stable
across versions.

On the other hand, using bundler 1 to parse a bundler 2 lockfile throws
an error if the "BUNDLED WITH" section in the lockfile is present.
Heroku explicitly removes the "BUNDLED WITH" section to workaround that,
but my end goal here is that heroku doesn't need to do that, and
instead respect the version of bundler the lockfile was created with
(the BUNDLED WITH section).

So, by not relying on bundler internals, we make sure that the removal
of the "BUNDLED WITH" section is no longer necessary.

This also makes the code consistent with other places that parse the
lockfile directly without bundler, like the code to detect the "BUNDLED
WITH" section and remove it.